Added comments and split features out
Added comments to required functions, updated log statements and
created service folder locations so that db does not get too cluttered
with the backend features.
