+++ apilayer/config/common.go
@@ -5,3 +5,5 @@ var BASE_LICENSE_KEY = []byte("2530d6a4-5d42-4758-b331-2fbbfed27bf9")
 
 const CTO_USER = "community_test"
 const CEO_USER = "ceo_test"
+
+const DefaultTokenValidityTime = "15"
